Added class to support lateral menu.
